Combine program and shader repository view models
